creating a simple calculators.
I've done this project as i a beginner in github and also to kill time and make the rest of the time i have in my hands worth it. Even though this is a simple project it will act as a stepping stones for my future projects and also helped me in gaining experience in github.thank you for this amazing opportunity.